Add test coverage for pipelinerun resolution edge case
We should never hit this case since the state should be caught when the
PipelineRun is created, but it seems weird to a) write the resoultion
code so that it could explode if the creation validation logic changes
and also b) weird to not cover this logic if we're going to add it.

I feel like this might be a slippery slope... We'll see!