Fix handling of case where non-existent resource is referenced
The helm end to end test was trying to bind a Pipeline’s resource to a resource that didn’t exist (my-special-resource). Debugging and fixing this revealed several problems and bugs which we fixed.

* If a PipelineRun didn’t bind the required resource, the reconcile function would continue on anyway (missing return statement!)
* If resolving a PipelineRun failed for any reason other than that a referenced Task couldn’t be find, the reconciler would pick up the PipelineRun again and continue trying to re-reconcile it. Now any resolution error will result in the PipelineRun being considered invalid and stopping.

The above all snuck in because @bobcatfish didn’t pay enough attention to covering the new functionality in integration tests at the reconcile level and has learned her lesson 😅

In the helm test itself, when the PipelineRun failed, we continued on, which led to tests taking much longer than needed to fail and also getting into a weird state (e.g. removing helm would fail because it was never set up). Now we stop the test immediately if the PipelineRun fails (though this could result in some desired cleanup in the cluster not occurring).

(It also looked like a test case at the reconciler level, which handled missing parameters, got wiped out at some point, so that’s back now.)

Co-authored-by: Nader Ziada <nziada@pivotal.io>