Add test coverage for pipeline providedBy
Adding an additional case for when a Task tries to depend on a Resource
from a previous Task that doesn't exist. Had previously covered this
only in the case where the first Task tries to express a `providedBy`,
but this is a special case b/c the first Task should never have any
`providedBy` constraints.