Rename `providedBy` to `from`
`providedBy` was a way for users to indicate when a Tasks's input should
come from the output of another Task - including any changes that Task
had made to it (e.g. adding files to a checkout of a git repo by
building a binary). We have struggled with an intuitive way to describe
it - orginally this was called `passedConstraint`. In the design doc for #320,
@sebgoa suggested we call this `from` which seems like the clearest name
yet!

Fixes #320 (remaining item in #320 to add `runAfter` will be moved to
another issue)