Fix flaky unit tests in resources package
Sorting the slices from `GetOutputSteps` and `GetInputSteps` in test,
to make sure we have a consistent ordering of the slice.

As maps have non-deterministic order in Go, and as we don't sort in
the actual methods (I don't think we need to have a deterministic
slice), the compared slice may not be in the same order. Sorting it by
name fixes that.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>