Update docs such that they can be used in knative/docs
In version 0.4 of knative, we'd like to release build-pipeline, which
means that the docs repo (https://github.com/knative/docs) should talk
about build-pipeline.

I think it's important to make sure that docs are updated with the code,
so I propose we have living versions of these docs in this repo, and
every time we release, we copy the docs into the knative/docs repo, so
that this repo has the correct docs at HEAD and knative/docs has the
correct docs as of the latest release. (How we want to deal with docs
for previous releases will have to be another story!)

In this PR, I've gone through [the knative docs for
build](https://github.com/knative/docs/tree/master/build) and shaped our
existing docs to be similar, which meant:

* Taking content out of using.md and splitting it into a file for each
  type of CRD
* Adding reference docs for the CRD fields (fixes #193)
* Most of concepts.md and some of using.md became the README file which
  lives at the top level of the `docs` dir here and hopefully the new
  `pipelines` dir in knative/docs

This also meant gaining a lot of docs by taking existing Build docs and
updating them to talk about pipelines (e.g. docs about auth)

Fixes #304