Add API compatibility policy
This doc will set out our backward incompatible change policy, so that
contributors have guidance on how to make changes, and so that users
will know what to expect.

In my original proposal, I wanted to make `Task` and `TaskRun` `beta`
immediately, but after some pushback from other contributors, we have
landed on leaving these `alpha` for a while (and having a hackathon to
get more user input!). (Which turned out to be good since we've already
found a couple of backwards incompatible changes we want to make!)

Fixes #34