Fix README and examples 😅
I tried to run the examples this morning and found that the docs were
wrong and also that the path in the example pipelines was now incorrect.
I can see how the docs could get out of sync but something odd is
happening with the pipeline paths b/c our tests should be catching that,
so I'll open a bug. In the meantime, these changes make it so we can at
least successfully create the pipelinerun (also tried the taskrun, which
works - I haven't tried the other examples).