Fix pvc vs bucket detection case 🌵
Depending on the `config-artifact-bucket` configMap, `build-pipeline`
will create either a PVC or GCS Bucket for sharing output between
task(run)s in the pipeline. There was a bug (or missing case) in the
code that checks if it needs to create PVC.

We want to create a PVC in the following cases:
- the `config-artifact-bucket` configMap is missing
- the `config-artifact-bucket` configMap data is empty
- the `config-artifact-bucket` configMap has no
`BucketLocationKey` (aka `location`) *or* it's value is empty.

The 2nd case wasn't correctly handled between
`InitializedArtifactStorage` and `GetArtifactStorage` functions.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>