Add new Resource type BuildGCS
In this commit, we added a new Resource Type "BuildGCS" which is
essentially mimics what GCSResource does in knative/build.
The BuildGCSResource object specifies
 - the GCS location of the artifact,
 - artifact type i.e. if the location points to Manifest file or an archive &
 - destination where it needs to be downloaded to.

The BuildGCSResource also uploads a source to GCS.
In future, We should combine "BuildGCS" and "GCS".
There are 2 avenues we can explore:

1. Find out what is the difference between images
"gcr.io/cloud-builders/gcs-uploader:latest" and
"override-with-gsutil-image:latest" image. If there isn't any
difference, we should use the
"gcr.io/cloud-builders/gcs-uploader:latest" image and get rid of
GCSResource

2. The GCSResource.GetDownloadContainerSpec should use the
"gcr.io/cloud-builders/gcs-fetcher:latest" instead of ""override-with-gsutil-image:latest"" and GCSResource.GetUploadContainerSpec
should use the "override-with-gsutil-image:latest" image.

I am more inclined towards option 1 to be more in-sync with
knative/build.

Co-authored-by:  Christie Wilson <bobcatfish@gmail.com>