Propagate labels from PipelineRun to TaskRun to Pods
Previously, labels were propagated from TaskRun to Build, but not from
Build to Pod. This PR fixes that, and also propagates labels from
PipelineRun to TaskRun, so that users are more easily able to filter,
identify, and query Pods created by Build Pipeline.

In all cases, if a user specifies a label whose key overlaps with one
of the labels that we set automatically, the user's label is ignored.

As mentioned in taskrun_test.go, the current tests do not fully
demonstrate that the TaskRun to Pod propagation works correctly since
they use Build as an intermediate form for comparison.

Fixes #458