Remove redundant label added to Pods
Previously, we added a label with key "build.knative.dev/buildName"
to Pods. That label was redundant with a different label whose key
is "pipeline.knative.dev/taskRun" as of beda1f87. Since we are no
longer depending on Knative Build at runtime, it seems best to
remove the redundant label.

Follow-up to #488.