Check labels in e2e PipelineRun tests and document label propagation
The e2e tests in test/pipelinerun.go now assert that custom labels set
on the PipelineRun are propagated through to the TaskRun and Pod, and
that the static labels that are added to all TaskRuns and Pods are
propagated correctly as well.

In addition, documentation has been added to explain that labels are
propagated and to mention the specific labels that are added
automatically to generated TaskRuns and Pods.

Follow-up to #488.