Build graph directly in `TestGetSchedulable`
Since we're going to be adding 2 ways to specify the graph (i.e. adding
`runAfter` in addition to `from`), updated this test to be agnostic to
how the graph was constructed.

Also updated the test to be table driven, to try to make it easier to
look at each case.