feat: add support for merging/batching multiple commits
This is a feature supported by Prow. Multiple 
branch:commits pairs can be specified in the env 
var `PULL_REFS`. The first pair is the 
branch:commit base, any subsequent pairs are 
commits to merge into the first pair. Prow uses 
this to batch up multiple commits for testing, and
also to test PRs which are opened against older
revisions of a branch against the `HEAD` of a 
branch.

This PR parses the `PULL_REFS` env var, and then
uses the git CLI to fetch any refs and then merge
them in.

Test coverage is about 80% - the success paths are
covered, error paths are not.