Update DAG to be more robust when handling multiple roots
While starting to add `runAfter` functionality, I wanted to update the
tests so that the tests for the building are separate from the tests
that get the next schedulable Task. While doing that I found what I
thought was a bug: if a graph has multiple roots, GetSchedulable would
sometimes (if there are no completed tasks in the other root ) only
return next tasks from one of the roots.

For example:

  b     a
  |    / \
  |   |   x
  |   | / |
  |   y   |
   \ /    z
    w

If you pass `x` to GetSchedulable, `b` wont' be returned, even though
theoretically this would be ready to run as well. Also, if a Task is
done, this implies everything before it is also done, even if it isn't
explicitly provided.

Eventually (a whole day later 😅) I realized that @vdemeester had
optimized by making the assumption that if anything beyond the roots was
"done", the roots themselves must already be executing, therefore we
were explicitly handling these two cases:

1. Nothing has started, so start all the roots
2. Otherwise look at everything, and if all its previous Tasks have
   completed, it must be ready to go.

I have changed this so that we instead walk the graph from the roots and
return the first schedulable nodes we encounter.

This is (only slightly) an improvement because:

1. If a root wasn't started successfully (i.e. if the controller itself
   encounters an error before starting one of the roots), we'll still
   try to start the next node on the next reconcile
2. If GetSchedulable is called with "done" Tasks that don't make sense
   (e.g. if we jump straight into the middle of the graph without
   completing the rest), we can catch that

This required adding `Next` pointers as well as `Prev` pointers - which
unfortunately meant we can't compare graphs with `cmp.Diff` because it
will have a stack overflow 😅