Rename project and API group to "Tekton Pipelines"
Instead of calling this "Pipeline CRD" or "Knative Pipelines", we'd like
to call this project "Tekton Pipelines". Before our first release, 0.1,
which we are intentionally decoupling from the Knative release process,
we want to change the api groups to reflect the new name.

The release scripts are hard-coded to use `knative` as the name, so we
will have to work around that for now.