Add release notes and submitter checklist to pull request template
Add a release notes section to each PR which can be scraped to create
each release. We may eventually want to create a
[changelog file](https://keepachangelog.com/en/0.3.0/) in the repo as
well but for now we can use the PRs and releases.

Since I was adding a pull request template anyway, I got started on a
submitter checklist as well, which we can iterate on and add more
specific items to as they come up.

For versioning, we're going to start with 0.1 and establish a regular
release schedule which we will probably iterate on. We will be following
semantic versioning (https://semver.org/).

Fixes #72