Work around markdown lint bug 😭
There is a bug in the markdown linter (mdl) that causes it to error out
if a link is bolded, e.g.:

```
**[I am a link](https://example.com)**
```

https://github.com/markdownlint/markdownlint/issues/211

Having the `<br/>` after the link seems to stop this from happening
(maybe that's why @a-roberts added it in the first place?). It wasn't
obvious why the `<br/>` was there so i n #473 I tried to remove it and
then had to spend some time debugging the cryptic ruby error.