Remove `logsURL` from step status
In #549 @hrishin pointed out that it's hard to understand from the step
status exactly which step did what. While looking at this I realized
that we have included a field `logsURL` which we never populate - I
thought this was copied over from Build but it was actually from our
original prototype API and we have never used it. In #107 we should be
revisiting making logs available and we may add in something like this,
but since we're not using it and it's not clear if we ever will, let's
remove it for now.