Build an image for executing ko 🏗️
Instead of doing a lot of hacks to make sure we have all the tools we
need to auth + invoke ko, let's build an image that has what we need in
advance.

Eventually we should be able to build this image and refer to the built
image in our steps (#639) but for now we'll have to hardcode it.

We may also improve this image in #631, or decide to move away from `ko`
entirely.