Add link to Tekton Pipelines Working Group calendar event 📱
We want to start having a 1 hr long working group meeting where we can
spend a significant amount of time reviewing designs and having
contributor discussions.

This PR adds a link to this working group calendar invite (which
hopefully works!) and also info abotu the tekton-dev@ group we'll use
for access to docs.