Add feature proposal guidelines 💡
Now that we've got more folks excited to contribute, we want to make
sure they have a smooth experience - it can be painful to put in the
effort to implement something only to have the entire design called into
question and to have to start over again. Hopefully these guidelines can
give contributors an idea of the best way to approach proposing changes.