Add `latest` tag to most recently released images 🔖
Looking again at the knative/test-infra release scripts, it looks like
they now tag `latest` on the most recently released images, which makes
sense for us as well (otherwise you can't pull without an explicit tag).

It also looks like the knative release process has changed such that the
process is to copy the most recent nightly-build, but we can decide if
that's the route we want to go down later.

Remove a comment about parsing and `ko`'s declarative model - in #631 we
can decide what we want to do about using `ko` for releasing in the
future.