Require `ko` >= 0.1 and simplify entrypoint copy container args
In #605, we required that `ko` was recent enough to be *guaranteed* to
have `/ko-app/{app}` (see google/go-containerregistry#380). As we now
point to `google/ko` and as we have a `ko` release (v0.1), we can
require this version for development.

Given the requirement above, we can simplify the `entrypoint` copy
container `args`.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>