Update release dogfood to work against prod 😌
The initial iteration of the Task and TaskRun worked just fine against
my own infrastructure, but when I tried to get it to work with our
production image registry and bucket, I struggled with getting the
various tools involved (kaniko, ko and gcloud) to all recognize the
service account.

In the end the easiest way to make this work was to lean on gcloud and
manually mounting the secrets into each Task. I still attach the secret
to the service account that runs the TaskRun because I think that's
needed for the bucket output uploading at the end.

The other goal here is to document this process so that someone else can
run the release Task next time :D (Though eventually we want to be able
to trigger this automatically to run in a different cluster, probably
`tekton-releases` own cluster.)

Two remaining unknowns that I'm hoping we can explore later:

- Could I have avoided using gcloud to auth by using servie accounts and
  secrets with annotations properly
  (https://github.com/tektoncd/pipeline/blob/master/docs/auth.md#guiding-credential-selection)
- Could I have avoided manually mounting the secrets by annotating the
  secret in some manner that would result in it being automatically
  mounted?