Wrap nop final step with entrypoint
`nop` image/step used to be executed *only* when all the previous step
were done sucessfully. As we are not using init containers anymore,
this is not true, `nop` will always be executed even in case of
previous failure (after the others though). This happens because we
don't use the `entrypoint` wrapper for this final step.

This fixes it by wrapping `nop` with entrypoint — just like other
steps. That way, it will only print `build successful` in case of
actual successful run 💃

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>