Fix a couple of small name issues in tutorial pipeline
- The tutorial pipeline names the input resources "workspace" and
"image" but the build-docker-image-from-git-source task expects
"docker-source" and "builtImage". The pathToContext also needed to
be updated to match the path that the git repo will be checked out at.

- The pipeline is configured to use the deploy-using-kubectl task and
provides an input git resource named "workspace" but the
deploy-using-kubectl task expects an input git resource named "source"
and the `path` param needed to be updated to reflect that.