Update CONTRIBUTING.md to link to guidelines in community repo 😎
Now that I've put so many guidelines in the community repo, instead of
duplicating them here, we can link to them! But we want to still make
sure they are easy for people to find - it can get very confusing to
have stuff spread across multiple repos so we should try to make it as
clear as possible where we've put everything!