Remove first WaitTaskRun call when creating TR pod
This WaitTaskRun call was moved in a commit but then reinstated the
following day. Looks like this being reinstated was an artifact of a
rebase for a separate PR and the removal of the call was the correct
thing to do.

The original change message for this removal is as follows:

> We can save one goroutines per pod creation by waiting for TaskRun
> only if the pod creation was successful.

This was the original PR with the call removal:
https://github.com/tektoncd/pipeline/pull/674/

And then it looks like it was reinstated here:
https://github.com/tektoncd/pipeline/pull/648/files#diff-b52b9f382ad878e6aa2a85921bb19065R287