Make first example of running e2e tests include longer timeout ⏱️
I foolishly tried to use the example at the top of the test README when
I was trying to run the timeout tests (for the #731 investigation) and I
was sad when it failed b/c the default timeout was too short. This is
the timeout we are using when running CI anyway!