Add docs to help ppl new to the project ramp up on the context 🏫
These docs and links come from @sbwsg! As he was ramping up he put
together this great list of resources (and the description wording is
mostly his too).

I hope we can continue to add to this ramp up section, definitely want
to make it as easy as possible for ppl to ramp up on the project, even
if they have no prior experience with k8s (let alone CRDs!).

Thanks @sbwsg :D <3