Fix TestStorageBucketPipelineRun to actually, y'know, run.
It was using the `UpdateConfigMap` function from Knative test, but
that assumes that `configMap.Data` is already a non-nil map. Since in
a normal test scenario with the default `config-artifact-bucket.yaml`
in `config/`, the relevant `configMap.Data` is in fact nil, the test
as written will just panic all the time.

Mind you, the test will still *fail*, but that'll be due to #819.

We haven't caught this, or #819, due to not running the bucket-related
tests in CI!

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>