Artifact bucket storage volume name should be consistent with mounts
The name of the volume added to the pod spec differed from the name of
the volume referenced in the container volume mount, for...well, no
good reason, as best as I could tell. The volume name used in the pod
spec also was using a generated random suffix, which just made things
more confusing. This change brings the naming in line with how we do
volume naming for GCS storage resource volumes, making the naming
consistent in both the pod spec and volume mount.

I also renamed the copy functions in `ArtifactStorageInterface`
because `GetCopyToContainerSpec` actually meant "copy *from* the
container to storage" and `GetCopyFromContainerSpec` meant "copy *to*
the container from storage", and that just always confused me at a
glance, so more verbose but clear names are nice to have.

And finally, I refactored `artifact_bucket_test.go` to make sure we're
actually testing that the volume name and volume mount name are the
same. Since that was the problem.

Fixes #819

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>