Deps: upgrade to 0.3.1 version of pipeline
This patch updates `pipeline` version to 0.3.1 and updates the vendor
directory.

The generation of `go.mod` took several attempts to get right and in
the end, it was generated by copying the Gopkg files and vendor from
pipeline repo, ensure the code complies and then run `go mod init`
again.

Signed-off-by: sthaha@redhat.com