Improve testabilty of cli commands
Previously, it was hard to test if a command like `list` works fine as
the command was tightly coupled to a concrete instance of Pipeline clientset
rendering the usage of a fake client for testing impossible.

Previous design also had a drawback that commands got added
magically (`func init`) and encouraged use of package level variables
in root command which the sub commands would depend on (leaking the
internals).

The new refactored code makes use of packages to add sub commands e.g.
`pipeline` package and does not make use of any package level
variables. This design also inverts the dependency between parent command
and child command such that the parent now explicitly need to add a
child command. The patch also makes it possible to test commands as the
commands make use of an interface (`Params`) to obtain the `Clientset`

Signed-off-by: sthaha@redhat.com