Storage end to end test can run without env var credentials ✨
The storage end to end test required that an environment variable be
configured that points to json credentials on disk which can be used to
pull from a private GCS bucket. These are not actually required to run
the test if the defualt credentials in the namespace have the
appropriate permissions - which is the case when the test runs in CI via
Prow!

With this change, the test will now be run in CI. This partially
addresses #821, but TestStorageBucketPipelineRun still requires this env
var to run.

Also removed the `read-secret-env` step from the storage test. If the
credential isn't present, the previous step in the Task would fail. Also
nothing is verifying the output of this step. (Maybe this was for
debugging?)

Co-authored-by: Christie Wilson <bobcatfish@gmail.com>