Graceful sidecar support
TL;DR Tekton Pipelines now gracefully start and stop with sidecars
running alongside TaskRun pods.

Long Version :-

TaskRun pods can have sidecars injected by Admission Controllers. This
is common practice with Istio where a proxy sidecar is injected into pods
so that they can be included as part of a service mesh. Unfortunately
there is no built-in Kubernetes mechanism to set the lifetime of a
sidecar container to match that of the pod's "primary" container. In
practice this means that pods injected with sidecars will not stop until
both the primary and all sidecar containers are stopped.

Prior to this commit injected sidecars would cause a TaskRun pod to run
forever, even when all Step containers of a Task had completed. This
commit introduces mechanisms to start and stop sidecars gracefully
meaning that a) Step containers wait until all sidecars are in a ready
state before starting and b) Sidecar containers are stopped when a
TaskRun's Step containers are done.

No end-to-end tests have been added as part of this PR because doing so
would require Istio to be enabled on the CI cluster (or a dummy
Admission Controller / Mutating Webhook thing to be written just for
injecting containers into test Pods). The Istio requirement would put an
undue burden on those users running tests in non-GKE environments. It's
currently planned for a follow-up PR to introduce a) user-defined
sidecars and b) a Tekton-injected sidecar that performs logging. Once that
work is complete it will be much simpler to e2e test this sidecar code.

Further to the above, a number of smaller refactors have taken place in
this PR:

- a sidecars package has been created to encapsulate the logic for
stopping sidecars with the nop container image
- the updateStatusFromPod func has been moved into a taskrun/status
package and broken down into constituent helpers
- a small amount of dead code has been eliminated

Co-authored-by: Jose Blas Camacho Taboada <joseblas@gmail.com>
Co-authored-by: Scott Seaward <sbws@google.com>