Proper OS exit status when a failure occures when listing
We were just doing fprintFing when there was a failure but not returning
properly as error to cobra.

The problem with this was that cobra would print the whole command usage and our message
would get lost along the way, so we use the SilenceUsage flag to silence
it.

cobra takes care of printing error message so we don't have to do it twice in
main.go too

Errors would still get printed when the user doesn't provide the right/needed
flags.

Closes: #22

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>