Remove OWNERS file from `hack` dir
In https://github.com/tektoncd/pipeline/pull/257 we made the knative
productivity working group owners of the hack dir in order to make it
easier for them to propogate changes across all the knative repos that
they worked on. Since then the project has moved out of the `knative`
org and also the scope of this dir has decreased (for example release
scripts no longer live here, we now have Tekton tasks for these).