Rename `containerTemplate` to `stepTemplate` but support both 👠
We used the term `containerTemplate` for this template b/c it is a
template of the containers used for each step - however we don't refer
to those as containers, we refer to them as steps, so I think calling
this `stepTemplate` makes slightly more sense and makes the relationship
between this template + the steps a bit more obvious.

This change is backward compatible b/c both `containerTemplate` and
`stepTemplate` will be supported until #977 where we will make the
backward incompatible change and remove `containerTemplate`.