Update tektoncd/plumbing vendor directory.
Fixes ./hack/update-deps.sh by pulling in fixes to library.sh.

https://github.com/tektoncd/plumbing/commit/1b17c1ce2b8362f80bb67aaa36be0677ea92ebd1#diff-5eee01526cae40785348322758638f44
introduced a bug that would attempt to use a script in a non-existent repo for
`./hack/update-deps.sh`. https://github.com/tektoncd/plumbing/pull/31 fixes
this to point back at the original repo. This updates the vendor directory to
include the changes to tektoncd/plumbing.