Add docs on how to access Run logs 🗎
As @cmoulliard pointed out, it's not obvious how to get to the logs for
a PipelineRun or a TaskRun. If you know how the underlying kubernetes
resources work you can figure it out but it can be hard to know where to
start. Plus, folks may not realize that we are working on better ways of
accessing logs.

And once we work on #107 we can build up these docs with more detail
about how to upload logs too.

Fixes #898