Restore the pod output in the build logs
A few E2E tests output the logs of the TaskRun pods to the build
log. This functionality is broken since the migration from
initContainers to containers.

Changing the `build_logs` module to loop over containers.
Include the name of containers since it's useful to map the log to
a specific step. Rename the function to `CollectPodLogs`.

Fixes #944