Adds PipelineTask.name metadata lables in TaskRun
If clients want to show the logs for particular TaskRun of a Pipeline,
it's not easy to relate which Task it belongs to in a Pipeline.
TaskRun should have some reference to PipelineTask.Name

In order to fix this issue, this patch adds a metadata.lable
in TaskRun while executing a Pipeline's Task.

Fixes
  - https://github.com/tektoncd/pipeline/issues/965