Initialize PipelineResources in declared order 📝
@pwittrock was trying to use PipelineResources such that the contents of
a GCS bucket could be copied over a previously checked out dir in a
github repo. Unfortunately, PipelineResources were being applied in the
opposite order that they were declared in!

This change makes it so that the PipelineResources will be applied in
the declared order, which means that the order these PipelineResources
are applied in is now part of our API.

Fixes #775

Co-authored-by: Eli Zucker <thezuck@google.com>