Add a release pipeline ⛏
This will need to be run manually for the first(s) releases, but it
sets pieces in place and help the dogfooding effort for Tekton
Pipelines.

It reuses golang tasks from the tektoncd/catalog *and* adds a
`goreleaser` Task (using goreleaser).

The pipeline runs the lint, then build the command and run the unit
tests. It finaly cross build the binary for all targeted architecture
and system and publish a draft release.

Although this should be considered still a work in progress, it is
still usable and will be run to make the v0.1.0 release of
tektoncd/cli.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>