Fix Pipeline error with ClusterTasks
It is possible to use `ClusterTask` in a Pipeline, for that you
specify that the `Task` you want to use is of type
`ClusterTask` (default being `Task`). But, before this change, the
`PipelineRun` using this `Pipeline` would fail as the `Kind`
information is not passed from the `PipelineTask.TaskRef` to the
actual `TaskRun`.

This fixes that by passing the `Kind` of a `TaskRef` when resolving
the `Task`/`TaskRun` in the `PipelineRun` reconcilier.

A yaml test is added as a regression test.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>